pyproject.toml
[project]
name = "best-hn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bitsandbytes>=0.44.1",
    "datasets>=3.0.1",
    "dicttoxml>=1.7.16",
    "flash-attn==2.6.3",
    "hf-transfer>=0.1.8",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "jupyter>=1.1.1",
    "liger-kernel>=0.3.1",
    "matplotlib>=3.9.2",
    "peft>=0.13.2",
    "polars>=1.9.0",
    "python-dotenv>=1.0.1",
    "sglang[all]>=0.3.3.post1",
    "torch==2.4.0",
    "tqdm>=4.66.5",
    "transformers>=4.45.2",
    "trl>=0.11.4",
    "wandb>=0.18.3",
]

[tool.uv.sources]
transformers = { git = "https://github.com/huggingface/transformers" }
sglang = { path = "deps/sglang/python", editable = true }
